pyproject.toml
[project]
name = "san-fransico-311-calls"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbt-core>=1.10.13",
    "dbt-postgres>=1.9.1",
    "duckdb>=1.4.0",
    "ipykernel>=6.30.1",
    "pandas>=2.3.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "sodapy>=2.2.0",
]
